Project 5 Continuation
Add navigation right bar button, when tapped an alert controller with a text field will appear asking for a answer.
Created 5 methods promptForAnswer, submit, isPossible, isOriginal, isReal.

promptForAnswer method is called when the user taps on the navigation right bar button item. Once answer is submited, submit method will be called
submit method takes a parameter of string from the text field from the alert controller runs through 4 if else statements to check if the word can be added. If not it will display what's wrong.
isPossible method takes the answer and iterates each character to the current word being asked to make an anagram of and return a boolean.
isOriginal method returns a boolean to checks the answer if it is already usedWords array
isReal method checks where the submited asnwer is a real word using UITextChecker class and returns a boolean valuee