Finalization of Project (Improvements)
isReal method has a condition where the word entered has to be greater the three characters otherwise it returns false
Refactored the AlertController in submit method and moved it to a new method called showErrorMessage when one of the conditions in the submit method are not met it will be called instead.
Added a LeftBarButtonItem where it calls startGame method to reset the game.